10x Better Decompiler.
Ok.

+++ me/alpha432/oyvey/util/RusherHackUtil.java
@@ -1,38 +1,46 @@
+/*
+ * Decompiled with CFR 0.151.
+ */
 package me.alpha432.oyvey.util;
 
 import java.lang.reflect.Field;
+import me.alpha432.oyvey.util.Util;
+import me.alpha432.oyvey.util.Wrapper;
 
-public class RusherHackUtil implements Util {
-  public static Field renderPosX;
-  
-  public static Field renderPosY;
-  
-  public static Field renderPosZ;
-  
-  public static double getRenderPosX() {
-    try {
-      return ((Double)renderPosX.get(Wrapper.mc.getRenderManager())).doubleValue();
-    } catch (Exception e) {
-      e.printStackTrace();
-      throw new IllegalStateException(e);
-    } 
-  }
-  
-  public static double getRenderPosY() {
-    try {
-      return ((Double)renderPosY.get(Wrapper.mc.getRenderManager())).doubleValue();
-    } catch (Exception e) {
-      e.printStackTrace();
-      throw new IllegalStateException(e);
-    } 
-  }
-  
-  public static double getRenderPosZ() {
-    try {
-      return ((Double)renderPosZ.get(Wrapper.mc.getRenderManager())).doubleValue();
-    } catch (Exception e) {
-      e.printStackTrace();
-      throw new IllegalStateException(e);
-    } 
-  }
+public class RusherHackUtil
+implements Util {
+    public static Field renderPosX;
+    public static Field renderPosY;
+    public static Field renderPosZ;
+
+    public static double getRenderPosX() {
+        try {
+            return (Double)renderPosX.get(Wrapper.mc.getRenderManager());
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+            throw new IllegalStateException(e);
+        }
+    }
+
+    public static double getRenderPosY() {
+        try {
+            return (Double)renderPosY.get(Wrapper.mc.getRenderManager());
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+            throw new IllegalStateException(e);
+        }
+    }
+
+    public static double getRenderPosZ() {
+        try {
+            return (Double)renderPosZ.get(Wrapper.mc.getRenderManager());
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+            throw new IllegalStateException(e);
+        }
+    }
 }
+
